feat(validator): strict mode
