+++ packages/validator/src/index.ts
@@ -1,1 +1,2 @@
 export * from './ui-validator';
+export type { UiValidatorErrors, UiValidatorData, UiValidatorSchema, UiValidatorResult } from './types';
