+++ packages/validator/src/types/schema-type.ts
@@ -6,7 +6,7 @@ export type UiValidatorExpectationRule<T> = {
   /** The expected value for this rule */
   expected: T;
   /** The error message to provide in the response if this check doesn't pass */
-  error?: UiValidatorError;
+  error: UiValidatorError;
 };
 
 /** The value for each rule */
@@ -16,7 +16,7 @@ export type UiValidatorLengthRule = {
   /** The expected max length for this rule */
   max: number;
   /** The error message to provide in the response if this check doesn't pass */
-  error?: UiValidatorError;
+  error: UiValidatorError;
 };
 
 /** The value for each rule */
@@ -26,7 +26,7 @@ export type UiValidatorRangeRule = {
   /** The expected max range for this rule */
   max: number;
   /** The error message to provide in the response if this check doesn't pass */
-  error?: UiValidatorError;
+  error: UiValidatorError;
 };
 
 /** Set of possible rules for each field */
