+++ packages/validator/src/ui-ruler.ts
@@ -3,63 +3,67 @@ import { UiValidatorRules } from './types';
 export class UiRuler {
   private rules: UiValidatorRules = {};
 
-  isRequired(errorMessage?: string, errorCode?: string | number): UiRuler {
+  isRequired(errorMessage?: string): UiRuler {
     this.rules.required = {
       expected: true,
-      error:
-        errorMessage || errorCode
-          ? {
-              message: errorMessage,
-              code: errorCode,
-            }
-          : undefined,
+      error: errorMessage
+        ? {
+            message: errorMessage,
+          }
+        : {
+            message: 'This is required',
+          },
     };
 
     return this;
   }
 
-  type(type: 'string' | 'numeric' | 'email' | 'phone', errorMessage?: string, errorCode?: string | number): UiRuler {
+  isOptional(): UiRuler {
+    return this;
+  }
+
+  type(type: 'string' | 'numeric' | 'email' | 'phone', errorMessage?: string): UiRuler {
     this.rules.type = {
       expected: type,
-      error:
-        errorMessage || errorCode
-          ? {
-              message: errorMessage,
-              code: errorCode,
-            }
-          : undefined,
+      error: errorMessage
+        ? {
+            message: errorMessage,
+          }
+        : {
+            message: `This is not a valid ${type}`,
+          },
     };
 
     return this;
   }
 
-  range(min: number, max: number, errorMessage?: string, errorCode?: string | number): UiRuler {
+  range(min: number, max: number, errorMessage?: string): UiRuler {
     this.rules.range = {
       min,
       max,
-      error:
-        errorMessage || errorCode
-          ? {
-              message: errorMessage,
-              code: errorCode,
-            }
-          : undefined,
+      error: errorMessage
+        ? {
+            message: errorMessage,
+          }
+        : {
+            message: `This is not in valid range of ${min} and ${max}`,
+          },
     };
 
     return this;
   }
 
-  length(min: number, max: number, errorMessage?: string, errorCode?: string | number): UiRuler {
+  length(min: number, max: number, errorMessage?: string): UiRuler {
     this.rules.length = {
       min,
       max,
-      error:
-        errorMessage || errorCode
-          ? {
-              message: errorMessage,
-              code: errorCode,
-            }
-          : undefined,
+      error: errorMessage
+        ? {
+            message: errorMessage,
+          }
+        : {
+            message: `This does not have a valid length of min ${min} and max ${max}`,
+          },
     };
 
     return this;
